requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
orjson==3.9.7
outcome==1.2.0
packaging==23.1
Pillow==10.0.1
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
win32-setctime==1.1.0
wsproto==1.2.0
